+++ common/src/main/java/io/github/mortuusars/horseman/mixin/HorsemanMixinPlugin.java
@@ -2,6 +2,7 @@ package io.github.mortuusars.horseman.mixin;
 
 import com.google.common.collect.ImmutableMap;
 import com.llamalad7.mixinextras.MixinExtrasBootstrap;
+import io.github.mortuusars.horseman.PlatformHelper;
 import org.objectweb.asm.tree.ClassNode;
 import org.spongepowered.asm.mixin.extensibility.IMixinConfigPlugin;
 import org.spongepowered.asm.mixin.extensibility.IMixinInfo;
@@ -13,7 +14,9 @@ import java.util.function.Supplier;
 
 public class HorsemanMixinPlugin implements IMixinConfigPlugin {
 
-    private static final Map<String, Supplier<Boolean>> CONDITIONS = ImmutableMap.of();
+    private static final Map<String, Supplier<Boolean>> CONDITIONS = ImmutableMap.of(
+            "io.github.mortuusars.horseman.mixin.ServerGamePacketListenerImplMixin", () -> !PlatformHelper.isModLoaded("horsebuff")
+    );
 
     @Override
     public void onLoad(String mixinPackage) {