+++ common/src/main/java/io/github/mortuusars/horseman/mixin/ServerGamePacketListenerImplMixin.java
@@ -24,7 +24,7 @@ public abstract class ServerGamePacketListenerImplMixin {
 //    }
 
     @ModifyConstant(method = "handleMoveVehicle", constant = @Constant(doubleValue = 0.0625))
-    private double onHandleMoveVehicle(double value){
+    private double onHandleMoveVehicle(double value) {
         if (player.getRootVehicle() instanceof AbstractHorse && Config.Common.FIX_HORSE_MOVED_WRONGLY.get())
             return 0.36;
         return value;